requirements.txt
Flask
gunicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
nltk
scikit-learn
pandas
